View Balance Completed

+++ capstone/tenmo-server/src/main/java/com/techelevator/tenmo/dao/AccountDao.java
@@ -6,18 +6,18 @@ import java.math.BigDecimal;
 import java.util.List;
 
 public interface AccountDao {
-    public BigDecimal getAccountBalance(int id);
+    public BigDecimal getAccountBalance(long id);
 
-    public Account getAccountById(int id);
+    public Account getAccountById(long id);
 
     public List<Account> getAllAccounts();
 
-    public Account getAccountByUserId(int id);
+    public Account getAccountByUserId(long id);
 
-    public BigDecimal addBalance(int id, BigDecimal amount);
+    public BigDecimal addBalance(long id, BigDecimal amount);
 
-    public BigDecimal subtractBalance(int id, BigDecimal amount);
+    public BigDecimal subtractBalance(long id, BigDecimal amount);
 
-    public boolean checkValidBalance(int id, BigDecimal amount);
+    public boolean checkValidBalance(long id, BigDecimal amount);
 
 }
